fixed inconsistency in settings args in plotting
plot_lightcurves was being passed the settings as a dict when it should have been passed the string that corresponded to the path to the settings_file. May be good to have greater consistency in which functions require the settings_file as a string and those that want the settings and model dicts.

In the case of something like fitting.generate_job(), the settings and model dicts are passed separately because the model_dict in this instance is one of the nested dictionary items in the overall models key in settings.json that corresponds to a given model. Could have discussion about whether it's better to have these functions execute on a model by model basis or if they should just get all the models and then generate/submit jobs for all models for a given object.

Co-Authored-By: bking-astro <71393384+bking-astro@users.noreply.github.com>